Struct rust-bt::tracker_udp::UdpTracker

pub struct UdpTracker {
    // some fields omitted
}

Methods

impl UdpTracker

fn new(url: &str, info_hash: &[u8]) -> IoResult<UdpTracker>

Creates a new UdpTracker object.

This function will send out a connect request on all available IPv4 interfaces and will use the interface that receives a response first.