Enum rust-bt::upnp::UPnPInterface

pub enum UPnPInterface {
    Root(String, StrPos, StrPos, StrPos, StrPos, StrPos),
    Device(String, StrPos, StrPos, StrPos, StrPos, StrPos),
    Service(String, StrPos, StrPos, StrPos, StrPos, StrPos),
    Identifier(String, StrPos, StrPos, StrPos, StrPos, StrPos),
}

Type used to represent different interfaces available on a network as defined by the UPnP specification. This type can be used to get very general information about an interface but also to get access to the services exposed by the interface.

Variants

Root

Device

Service

Identifier

Methods

impl UPnPInterface

fn find_all(from_addr: SocketAddr) -> IoResult<Vec<UPnPInterface>>

Sends out a search request for all UPnP interfaces on the specified from_addr.

This is a blocking operation.

fn find_services(from_addr: SocketAddr, name: &str, version: &str) -> IoResult<Vec<UPnPInterface>>

Sends out a search request for all UPnP services on the specified from_addr.

This function does not check to make sure that only services responded to our request, so it is assumed that devices on the network correctly follow the UPnP specification. As of now, any services that are not defined at http://upnp.org/sdcps-and-certification/standards/sdcps/ will not be found by this method (use find_all and filter with is_service).

This is a blocking operation.

fn find_devices(from_addr: SocketAddr, name: &str, version: &str) -> IoResult<Vec<UPnPInterface>>

Sends out a search request for all UPnP devices on the specified from_addr.

This function does not check to make sure that only devices responded to our request, so it is assumed that devices on the network correctly follow the UPnP specification. As of now, any devices that are not defined at http://upnp.org/sdcps-and-certification/standards/sdcps/ will not be found by this method (use find_all and filter with is_device).

This is a blocking operation.

fn find_uuid(from_addr: SocketAddr, uuid: &str) -> IoResult<Option<UPnPInterface>>

Sends out a search request for a specific UPnP interface on the specified from_addr.

This function returns one interface but does not check to see if more than one device responded. If more then one device responded (which should not happen) then the first response is what will be returned.

This is a blocking operation.

fn is_root(&self) -> bool

Check if the current UPnPInterface represents a root device on the network.

fn is_device(&self) -> bool

Check if the current UPnPInterface represents a device on the network.

Note that this function will return false if the type is a root device.

fn is_service(&self) -> bool

Check if the current UPnPInterface represents a service on the network.

fn is_identifier(&self) -> bool

Check if the current UPnPInterface represents an identifier (uuid) on the network.

fn name<'a>(&'a self) -> &'a str

Get the device name, service specification, or uuid that this interface has identified with.

fn version<'a>(&'a self) -> &'a str

Get the device version, service version, or no version (identifiers/root devices) that this interface has identified with.

fn location<'a>(&'a self) -> &'a str

Get the location of the root device description page corresponding to this UPnPInterface.

fn st<'a>(&'a self) -> &'a str

Get the search target.

fn usn<'a>(&'a self) -> &'a str

Get the full Unique Service Name.

fn service_desc(&self) -> IoResult<ServiceDesc>

Get the service description for this service.

Note that if this is called on an object that is not service, an error will always be returned.

This is a blocking operation.