Struct rust-bt::upnp::ServiceDesc

pub struct ServiceDesc {
    // some fields omitted
}

Type used to represent a service description. Service actions can be viewed as well as related in and out parameters and service state variables.

Methods

impl ServiceDesc

fn parse(location: &str, st: &str) -> IoResult<ServiceDesc>

Takes a location corresponding to the root device description page of a UPnPInterface as well as the st field of the interface which will always be unique to a service/device.

This is a blocking operation.

fn soap_request(&self, action: &str, params: &[(&str, &str)]) -> IoResult<()>