Module rust-bt::util

Enums

Transport

Functions

gen_peer_id
Generates a peer id from a base identifier followed by random characters.

get_error
Used to fill an IoError with a kind and desc, leaving detail empty.

get_net_addrs
Returns a list of all local IPv4 Addresses.

get_path
Returns the path portion of a supplied url.

get_sockaddr
Returns the first found DNS entry as a SocketAddr for the specified url.

get_transport
Takes a url and returns the transport type that it specifies.

get_udp_sock
Attempts to open a udp connection on addr.

If the connection is unsuccessful, it will try again up to (attempts - 1) times, incrementing the port for each attempt.

get_udp_wait
The standard wait algorithm defined in the UDP Tracker Protocol.